pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "latentlens"
version = "0.1.0"
description = "LatentLens: Interpreting Visual Tokens in Vision-Language Models"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Benno Krojer", email = "benno.krojer@mila.quebec" },
    { name = "Shravan Nayak" },
    { name = "Oscar Manas" },
    { name = "Vaibhav Adlakha" },
    { name = "Desmond Elliott" },
    { name = "Siva Reddy" },
    { name = "Marius Mosbach" },
]
keywords = [
    "vision-language-models",
    "interpretability",
    "multimodal",
    "transformers",
    "deep-learning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.3.0,<2.5.0",
    "torchvision>=0.18.0",
    "transformers>=4.45.0",
    "huggingface-hub>=0.25.0",
    "numpy>=1.26.0,<2.0.0",
    "pillow>=10.0.0",
    "einops>=0.8.0",
    "omegaconf>=2.3.0",
    "safetensors>=0.4.0",
    "ml-dtypes>=0.4.0",
    "accelerate>=1.0.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
train = [
    "wandb>=0.18.0",
]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]
all = [
    "latentlens[train,dev]",
]

[project.urls]
Homepage = "https://github.com/McGill-NLP/latentlens"
Documentation = "https://github.com/McGill-NLP/latentlens"
Repository = "https://github.com/McGill-NLP/latentlens"

[project.scripts]
latentlens = "molmo.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["latentlens", "molmo"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
